timestamping: do not store timestamp in packet
RFC 8219 requires to take a fresh timestamp right
after the sending of the frame, not after inserting
it into the queue.

To match a received packet against the saved timestamp
we store the index into the timestamp array into the
packet, together with the queue number from which the
packet was sent, the index + queue number uniquelly
identify a packet within the stream.